+++ src/nr_openai_observability/stream_patcher.py
